third_party/WebKit/LayoutTests/external/wpt/webxr/xrSession_requestAnimationFrame_getDevicePose.https-expected.txt

This is a testharness.js-based test.
PASS XRFrame getDevicePose updates on the next frame for immersive sessions
FAIL XRFrame getDevicePose updates on the next frame for non-immersive sessions assert_equals: expected null but got object "[object XRDevicePose]"
Harness: the test ran to completion.
